requirements.txt
torch>=2.1.0
torchvision>=0.16.0
diffusers>=0.24.0
transformers>=4.30.0
accelerate>=0.20.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
bitsandbytes>=0.41.0
safetensors>=0.3.1
omegaconf>=2.3.0
gradio>=4.0.0
psutil>=5.9.0
modal>=0.63.0
leonardo-ai-sdk>=1.0.0
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
google-genai>=0.3.0
